+++ src/db/DomainsQuery.php
@@ -9,9 +9,12 @@
 namespace flipbox\domains\db;
 
 use craft\db\QueryAbortedException;
+use craft\helpers\Db;
 use flipbox\craft\sortable\associations\db\SortableAssociationQuery;
 use flipbox\craft\sortable\associations\db\traits\SiteAttribute;
 use flipbox\domains\records\Domain;
+use flipbox\ember\db\traits\ElementAttribute;
+use flipbox\ember\helpers\QueryHelper;
 
 /**
  * @method Domain[] getCachedResult()
@@ -19,6 +22,8 @@ use flipbox\domains\records\Domain;
 class DomainsQuery extends SortableAssociationQuery
 {
     use traits\Attributes,
+        traits\FieldAttribute,
+        ElementAttribute,
         SiteAttribute;
 
     /**
@@ -30,6 +35,20 @@ class DomainsQuery extends SortableAssociationQuery
     }
 
     /**
+     * @param array $config
+     * @return $this
+     */
+    public function configure(array $config)
+    {
+        QueryHelper::configure(
+            $this,
+            $config
+        );
+
+        return $this;
+    }
+
+    /**
      * @inheritdoc
      *
      * @throws QueryAbortedException if it can be determined that there won’t be any results
@@ -37,14 +56,20 @@ class DomainsQuery extends SortableAssociationQuery
     public function prepare($builder)
     {
         // Is the query already doomed?
-        if (($this->fieldId !== null && empty($this->fieldId)) ||
-            ($this->domain !== null && empty($this->domain))
+        if (($this->field !== null && empty($this->field)) ||
+            ($this->domain !== null && empty($this->domain)) ||
+            ($this->element !== null && empty($this->element))
         ) {
             throw new QueryAbortedException();
         }
 
-        $this->applySiteConditions();
         $this->applyConditions();
+        $this->applySiteConditions();
+        $this->applyFieldConditions();
+
+        if ($this->element !== null) {
+            $this->andWhere(Db::parseParam('elementId', $this->parseElementValue($this->element)));
+        }
 
         return parent::prepare($builder);
     }